// apps/web/src/api/__generated__/types/CreateRecruitmentPostRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { CreateRecruitmentPostRequestContent } from './CreateRecruitmentPostRequestContent';
import type { CreateRecruitmentPostRequestEmploymentType } from './CreateRecruitmentPostRequestEmploymentType';
import type { CreateRecruitmentPostRequestExperienceLevel } from './CreateRecruitmentPostRequestExperienceLevel';
import type { CreateRecruitmentPostRequestReferencedPostIdsItem } from './CreateRecruitmentPostRequestReferencedPostIdsItem';
import type { CreateRecruitmentPostRequestTagsItem } from './CreateRecruitmentPostRequestTagsItem';
import type { CreateRecruitmentPostRequestWorkMode } from './CreateRecruitmentPostRequestWorkMode';

export interface CreateRecruitmentPostRequest {
  /**
   * 커버 이미지 미디어 ID
   * @nullable
   */
  coverMediaId?: string | null;
  /** 경력 수준 */
  experienceLevel: CreateRecruitmentPostRequestExperienceLevel;
  /** 근무 방식 */
  workMode: CreateRecruitmentPostRequestWorkMode;
  /** 고용 형태 */
  employmentType: CreateRecruitmentPostRequestEmploymentType;
  /**
   * 근무 지역
   * @nullable
   */
  location?: string | null;
  /**
   * 참조 게시글 ID 목록
   * @nullable
   */
  referencedPostIds?:
    | CreateRecruitmentPostRequestReferencedPostIdsItem[]
    | null;
  /** 모집 제목 */
  title: string;
  /**
   * 모집 마감 시각
   * @nullable
   */
  closesAt?: string | null;
  /** 모집 내용 */
  content: CreateRecruitmentPostRequestContent;
  /**
   * 기술 태그 목록
   * @nullable
   */
  tags?: CreateRecruitmentPostRequestTagsItem[] | null;
}